+++ DefenseApplet/src/application/HarmCalculator.java
@@ -4,27 +4,27 @@ import factors.IFactorsModel;
 
 public class HarmCalculator implements IHarmCalculator {
 
-	private static final int PHYS_DAMAGE_BONUS = 2;
-	private static final int SPEC_DAMAGE_BONUS = 2;
-	private int incomingHarm;
+	private static final long PHYS_DAMAGE_BONUS = 2;
+	private static final long SPEC_DAMAGE_BONUS = 2;
+	private long incomingHarm;
 
 	public HarmCalculator() {
 		this(320844000);
 	}
 
-	public HarmCalculator(int incomingHarm) {
+	public HarmCalculator(long incomingHarm) {
 		this.incomingHarm = incomingHarm;
 	}
 
 	@Override
-	public int calculate(PokemonStats pokemonStats, IFactorsModel factorsModel) {
-		int initialPhysicalDamage = (incomingHarm + PHYS_DAMAGE_BONUS) * factorsModel.getDefDenominator();
-		int initialSpecialDamage = (incomingHarm + SPEC_DAMAGE_BONUS) * factorsModel.getSpDefDenominator();
+	public long calculate(PokemonStats pokemonStats, IFactorsModel factorsModel) {
+		long initialPhysicalDamage = (incomingHarm + PHYS_DAMAGE_BONUS) * factorsModel.getDefDenominator();
+		long initialSpecialDamage = (incomingHarm + SPEC_DAMAGE_BONUS) * factorsModel.getSpDefDenominator();
 
-		int physicalDamage = initialPhysicalDamage / pokemonStats.getDef() / factorsModel.getDefNumerator();
-		int specialDamage = initialSpecialDamage / pokemonStats.getSpDef() / factorsModel.getSpDefNumerator();
+		long physicalDamage = initialPhysicalDamage / pokemonStats.getDef() / factorsModel.getDefNumerator();
+		long specialDamage = initialSpecialDamage / pokemonStats.getSpDef() / factorsModel.getSpDefNumerator();
 
-		int result = physicalDamage + specialDamage;
+		long result = physicalDamage + specialDamage;
 		result *= factorsModel.getHPDenominator();
 		result /= factorsModel.getHPNumerator();
 		return result / pokemonStats.getHP();
