+++ DefenseApplet/src/application/IHarmCalculator.java
@@ -4,6 +4,6 @@ import factors.IFactorsModel;
 
 public interface IHarmCalculator {
 
-	int calculate(PokemonStats pokemonStatsBetter, IFactorsModel factorsModel);
+	long calculate(PokemonStats pokemonStatsBetter, IFactorsModel factorsModel);
 
 }
